-l option

+++ include/ft_ping.h
@@ -6,7 +6,7 @@
 /*   By: mamartin <mamartin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/10 17:46:56 by mamartin          #+#    #+#             */
-/*   Updated: 2022/09/16 23:09:53 by mamartin         ###   ########.fr       */
+/*   Updated: 2022/09/16 23:31:27 by mamartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ typedef struct s_options
 	uint8_t quiet;
 	int64_t count;
 	int interval;
-	int preload;
+	uint16_t preload;
 	int16_t ttl;
 	int32_t deadline;
 	int	pattern;
